fix(#4104): carry the SerializationHeader refusal's MESSAGE to the user (blocker A)
`HeaderSchemaError::Refused(e)` was logged with `tracing::error!` and then
DISCARDED: the function returned a bare `nom::Err::Error(.. ErrorKind::Verify)`,
which downstream became `UnsupportedFormat("… {:?}", nom_err)` and then
`Corruption("Failed to parse Statistics.db with enhanced parser: …")`. So opening
an SSTable whose header spells `FrozenType(Int32Type)` reported
`Corruption: … Error { input: [0x…], code: Verify }` — shape-identical to a
truncated or garbled file — while the precise, column-naming, citation-carrying
message the gate built survived only if `RUST_LOG` happened to be on.

The typed channel now runs all the way out:
- `parse_minimal_encoding_stats` / `parse_encoding_stats_fallback` return
  `Result<_, HeaderSchemaError>`; both key-type gates carry their `Error` too.
- `parse_nb_format_statistics_data_with_toc` decides on the VARIANT (never on
  message text) and propagates a `Refused`'s `Error` verbatim, keeping its kind;
  a `Structural` failure keeps the old wording and kind exactly.
- `parse_enhanced_statistics_file_detailed` /
  `parse_statistics_with_fallback_detailed` are the typed twins of the two nom
  entry points, whose signatures (and every existing caller) are unchanged; the
  flattening to `ErrorKind::Verify` now happens only in the nom wrapper.
- `StatisticsReader::open` uses the typed twin and returns a `Schema` refusal
  unchanged, relabelling only structural failures `Corruption`. `open()` still
  aborts either way (`component_loading` is untouched, and propagates a
  non-`Corruption` error unchanged by its own documented rule).

Pinned by `frozen_scalar_header_is_refused_and_never_marker_searched`, which now
asserts BOTH the variant and that the surfaced text names the refused type
verbatim, the writer rule and the `CQL3Type.java:647-651` oracle.

Also extracted `serialization_header/mod.rs`'s 400-line inline test module to
`serialization_header_tests.rs` (#1116/#1135): that file was ~1060 lines, so the
fail-closed marker-search change coming next cannot land in it otherwise.
