+++ cqlite-core/src/storage/sstable/statistics_reader.rs
@@ -5,7 +5,7 @@
 
 use crate::{
     error::{Error, Result},
-    parser::enhanced_statistics_parser::parse_statistics_with_fallback,
+    parser::enhanced_statistics_parser::parse_statistics_with_fallback_detailed,
     parser::statistics::{SSTableStatistics, StatisticsAnalyzer, StatisticsSummary},
     platform::Platform,
 };
@@ -92,13 +92,17 @@ impl StatisticsReader {
         let mut buffer = Vec::new();
         file.read_to_end(&mut buffer).await?;
 
-        let statistics = match parse_statistics_with_fallback(&buffer, gates.as_ref()) {
+        // `_detailed` keeps the TYPED error, so a SEMANTIC refusal keeps its own
+        // kind and its column-naming message instead of becoming `code: Verify`
+        // relabelled `Corruption` (#4104; `open()` aborts either way). Decided on
+        // the VARIANT, never on message text — rationale on the parser's doc.
+        let statistics = match parse_statistics_with_fallback_detailed(&buffer, gates.as_ref()) {
             Ok((_, stats)) => stats,
+            Err(e @ Error::Schema(_)) => return Err(e),
             Err(e) => {
                 return Err(Error::corruption(format!(
-                    "Failed to parse Statistics.db with enhanced parser: {:?}",
-                    e
-                )));
+                    "Failed to parse Statistics.db with enhanced parser: {e}"
+                )))
             }
         };
 
@@ -108,16 +112,11 @@ impl StatisticsReader {
         // The checksum field (header.checksum) contains a value but we cannot validate it
         // without knowing the exact algorithm Cassandra 5.0+ uses for this file format.
         //
-        // Known limitations:
-        // - Files with corrupt data may be silently accepted
-        // - No data integrity guarantee for Statistics.db parsing
-        // - M2 milestone will implement proper CRC32/Adler32/other validation
-        //
-        // This limitation is explicitly documented rather than silently ignored.
-        // Statistics.db is metadata-only (not critical path) so risk is acceptable for M1.
-        if statistics.header.checksum != 0 {
-            // Checksum present but not validated - this is a known M1 limitation
-        }
+        // Known limitations: corrupt data may be silently accepted, there is no
+        // integrity guarantee for Statistics.db parsing, and M2 will implement the
+        // real CRC32/Adler32 validation. Documented rather than silently ignored;
+        // Statistics.db is metadata-only, so the M1 risk is acceptable. (An empty
+        // `if header.checksum != 0 {}` used to stand here and executed nothing.)
 
         Ok(Self {
             file_path: path.to_path_buf(),
